requirements.txt
flask
python-dotenv
openai
requests
waitress
faiss-cpu
langchain-core
langchain-community
langchain-openai
